# docs/swagger.yaml
basePath: /
definitions:
...
  requests.AuthorPayload:
    properties:
      name:
        type: string
    type: object
  requests.AuthorUpdatePayload:
    properties:
      name:
        type: string
    type: object
  requests.BookPayload:
    properties:
      author_id:
        type: integer
      barcode:
        type: string
      fixture_no:
        type: integer
      is_donation:
        type: boolean
      language:
        type: string
      level:
        description: ilkokul, ortaokul, ortak
        type: string
      name:
        type: string
      page_count:
        type: integer
      publisher_id:
        type: integer
      shelf_code:
        type: string
    type: object
  requests.BookUpdatePayload:
    properties:
      author_id:
        type: integer
      barcode:
        type: string
      fixture_no:
        type: integer
      is_donation:
        type: boolean
      language:
        type: string
      level:
        type: string
      name:
        type: string
      page_count:
        type: integer
      publisher_id:
        type: integer
      shelf_code:
        type: string
    type: object
...
  resources.AuthorCollection:
    properties:
      data:
        items:
          $ref: '#/definitions/resources.AuthorResource'
        type: array
      meta:
        $ref: '#/definitions/resources.PaginationMeta'
    type: object
  resources.AuthorResource:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
  resources.BookCollection:
    properties:
      data:
        items:
          $ref: '#/definitions/resources.BookResource'
        type: array
      meta:
        $ref: '#/definitions/resources.PaginationMeta'
    type: object
  resources.BookResource:
    properties:
      author:
        $ref: '#/definitions/resources.AuthorResource'
      author_id:
        type: integer
      barcode:
        type: string
      created_at:
        type: string
      fixture_no:
        type: integer
      id:
        type: integer
      is_currently_loaned:
        type: boolean
      is_donation:
        type: boolean
      language:
        type: string
      level:
        type: string
      name:
        type: string
      page_count:
        type: integer
      publisher:
        $ref: '#/definitions/resources.PublisherResource'
      publisher_id:
        type: integer
      shelf_code:
        type: string
      updated_at:
        type: string
    type: object
...
  resources.PaginationMeta:
    properties:
      current_page:
        type: integer
      from:
        type: integer
      last_page:
        type: integer
      per_page:
        type: integer
      to:
        type: integer
      total:
        type: integer
    type: object
...
  resources.PublisherResource:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  responses.ErrorBody:
    properties:
      details: {}
      message:
        type: string
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact: {}
  title: Safa API
  version: "1.0"
paths:
...
  /authors:
    get:
      description: Get a paginated list of Author records
      parameters:
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Items per page (default 15)
        in: query
        name: per_page
        type: integer
      - description: Search by name
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resources.AuthorCollection'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.ErrorBody'
      security:
      - BearerAuth: []
      summary: List all Author records
      tags:
      - Author
    post:
      consumes:
      - application/json
      parameters:
      - description: Author payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/requests.AuthorPayload'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/resources.AuthorResource'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.ErrorBody'
      security:
      - BearerAuth: []
      summary: Create an Author
      tags:
      - Author
  /authors/{id}:
    delete:
      parameters:
      - description: Author ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.ErrorBody'
      security:
      - BearerAuth: []
      summary: Delete an Author
      tags:
      - Author
    get:
      parameters:
      - description: Author ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resources.AuthorResource'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.ErrorBody'
      security:
      - BearerAuth: []
      summary: Get an Author by ID
      tags:
      - Author
    put:
      consumes:
      - application/json
      parameters:
      - description: Author ID
        in: path
        name: id
        required: true
        type: integer
      - description: Author update payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/requests.AuthorUpdatePayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resources.AuthorResource'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.ErrorBody'
      security:
      - BearerAuth: []
      summary: Update an Author
      tags:
      - Author
  /books:
    get:
      description: 'Get a paginated list of books with optional filters. Default sort:
        fixture_no ASC.'
      parameters:
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Items per page (default 15)
        in: query
        name: per_page
        type: integer
      - description: Filter by name (partial match)
        in: query
        name: name
        type: string
      - description: Filter by fixture number (exact)
        in: query
        name: fixture_no
        type: integer
      - description: Search by name or barcode
        in: query
        name: search
        type: string
      - description: Filter by author ID
        in: query
        name: author_id
        type: integer
      - description: Filter by publisher ID
        in: query
        name: publisher_id
        type: integer
      - description: Filter by level (ilkokul/ortaokul/ortak)
        in: query
        name: level
        type: string
      - description: Filter by donation flag
        in: query
        name: is_donation
        type: boolean
      - description: Sort column (e.g. fixture_no ASC)
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resources.BookCollection'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.ErrorBody'
      security:
      - BearerAuth: []
      summary: List all Book records
      tags:
      - Book
    post:
      consumes:
      - application/json
      parameters:
      - description: Book payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/requests.BookPayload'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/resources.BookResource'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.ErrorBody'
      security:
      - BearerAuth: []
      summary: Create a Book
      tags:
      - Book
  /books/{id}:
    delete:
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.ErrorBody'
      security:
      - BearerAuth: []
      summary: Delete a Book
      tags:
      - Book
    get:
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resources.BookResource'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.ErrorBody'
      security:
      - BearerAuth: []
      summary: Get a Book by ID
      tags:
      - Book
    put:
      consumes:
      - application/json
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: integer
      - description: Book update payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/requests.BookUpdatePayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resources.BookResource'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.ErrorBody'
      security:
      - BearerAuth: []
      summary: Update a Book
      tags:
      - Book
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"